op-challenger: Consolidate fault dispute contract code.
